Separate Executor to an independent class
