+++ plugins/sharding/sources.am
@@ -12,4 +12,5 @@ sharding_scripts =				\
 	dynamic_columns.rb			\
 	window.rb			\
 	stream_execute_context.rb		\
-	stream_shard_executor.rb
+	stream_shard_executor.rb		\
+	stream_executor.rb
